Move namespaces
