+++ src/Core/src/Platform/iOS/ViewExtensions.cs
@@ -866,5 +866,11 @@ namespace Microsoft.Maui.Platform
 
 			return null;
 		}
+
+		internal static bool HideSoftInput(this UIView inputView) => inputView.ResignFirstResponder();
+
+		internal static bool ShowSoftInput(this UIView inputView) => inputView.BecomeFirstResponder();
+
+		internal static bool IsSoftInputShowing(this UIView inputView) => inputView.IsFirstResponder;
 	}
 }
